# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "DeFi 资产管理系统 API 文档",
        "title": "Rotki Demo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/wallets": {
            "get": {
                "description": "获取所有钱包列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "获取钱包列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_rotki-demo_internal_models.Wallet"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的钱包",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "创建钱包",
                "parameters": [
                    {
                        "description": "钱包信息",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler.CreateWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.Wallet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wallets/{id}": {
            "get": {
                "description": "根据 ID 获取钱包详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "获取钱包",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "钱包 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.Wallet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "更新钱包信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "更新钱包",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "钱包 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "钱包信息",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler.UpdateWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.Wallet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "根据 ID 删除钱包",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "删除钱包",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "钱包 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_rotki-demo_internal_models.Address": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "asset_snapshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rotki-demo_internal_models.AssetSnapshot"
                    }
                },
                "chain_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "last_synced_at": {
                    "type": "string"
                },
                "tags": {
                    "description": "用户定义的标签",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rotki-demo_internal_models.Token"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "wallet": {
                    "description": "关系",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.Wallet"
                        }
                    ]
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_rotki-demo_internal_models.AssetSnapshot": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "data_source": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "raw_data": {
                    "$ref": "#/definitions/github_com_rotki-demo_internal_models.JSONMap"
                },
                "snapshot_time": {
                    "type": "string"
                },
                "total_usd_value": {
                    "type": "number"
                }
            }
        },
        "github_com_rotki-demo_internal_models.Chain": {
            "type": "object",
            "properties": {
                "chain_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "native_token_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_rotki-demo_internal_models.JSONMap": {
            "type": "object",
            "additionalProperties": true
        },
...
        "github_com_rotki-demo_internal_models.Token": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "关系",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.Address"
                        }
                    ]
                },
                "address_id": {
                    "type": "integer"
                },
                "balance": {
                    "type": "string"
                },
                "chain": {
                    "$ref": "#/definitions/github_com_rotki-demo_internal_models.Chain"
                },
                "chain_id": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_updated": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "usd_value": {
                    "type": "number"
                }
            }
        },
        "github_com_rotki-demo_internal_models.Wallet": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rotki-demo_internal_models.Address"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled_chains": {
                    "description": "启用的链 ID 列表，空表示所有链",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "description": "用户定义的标签",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handler.CreateWalletRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "enabled_chains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_api_handler.UpdateWalletRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "enabled_chains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}